Fix : resolve circular import deadlock
